Now handles classes with the same name in different packages.
The UI is also improved a bit. Interfaces are typeset in italics
(like in normal javadoc) and there is a MORE button that allows
you to explore search results further if you want. The size of
the results lists is doubled every time the button MORE is
used. On my computer it becomes slugish if you press MORE for
about 10 times (and the list of results is huge, of course).