Code refactoring, wildcards, incremental search.
The purpose of the refactoring is to make it easier to reuse code
when we'll search thru recent classes/packages as well as all
classes/packages.

The Finder code is isolated in a different class. The implementation
is (still) simple linear search using regular expressions. Seems
fast enough. The interface is suited for "incremental" search, that is,
you can search for up to X results and Finder will only spend time
necessary to find X results, not more. You can continue the search later.

The wildcard mechanis described in
  http://code.google.com/p/javadoc-reader/wiki/SearchString
is implemented.